% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spCEH.R
\docType{data}
\name{crs_Ire}
\alias{crs_Ire}
\title{CRS object for the TM75 Irish Grid projection.}
\format{
An sp CRS object
}
\source{
\url{https://spatialreference.org/}
}
\usage{
crs_Ire
}
\description{
A Coordinate Reference System object for the Transverse Mercator projection used by the Ordnance Survey in Ireland (TM75 / EPSG:29903)
}
\keyword{datasets}
