% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spCEH.R
\docType{data}
\name{crs_lonlat}
\alias{crs_lonlat}
\title{CRS object for null projection (longitude-latitude).}
\format{
An sp CRS object
}
\source{
\url{https://spatialreference.org/}
}
\usage{
crs_lonlat
}
\description{
A Coordinate Reference System object for the WGS84 lon-lat coordinate system / EPSG:4326
}
\keyword{datasets}
