% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spCEH.R
\docType{data}
\name{crs_OSGB}
\alias{crs_OSGB}
\title{CRS object for the OSGB projection.}
\format{
An sp CRS object
}
\source{
\url{https://spatialreference.org/}
}
\usage{
crs_OSGB
}
\description{
A Coordinate Reference System object for the Transverse Mercator projection used by the Ordnance Survey in Great Britain (OSGB / EPSG:27700)
}
\keyword{datasets}
