% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spCEH.R
\name{get_r_alt}
\alias{get_r_alt}
\title{Function to export UK altitude}
\format{
A SpatRaster (Raster) object from the terra package. Units: metres above mean sea level. CRS: EPSG 27700 (British National Grid)
}
\source{
\url{https://SRTM space shuttle terrain mission I think/}
}
\usage{
get_r_alt()
}
\description{
A raster layer of altitude in the UK. Object 'r_alt' returned to global environment.
}
\examples{
get_r_alt()
}
