% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spCEH.R
\name{get_r_Csoil}
\alias{get_r_Csoil}
\title{Function to export UK soil carbon}
\format{
A SpatRaster object from the terra package. Units: kg C / m2. CRS: EPSG 27700 (British National Grid)
}
\source{
\url{https://Bradley et al 2005/}
}
\usage{
get_r_Csoil()
}
\description{
A raster layer of soil carbon in the UK
}
\examples{
r_Csoil <- get_r_Csoil()
}
