% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spCEH.R
\name{get_r_lcm}
\alias{get_r_lcm}
\title{Function to export UK Land Cover Map classes}
\format{
A SpatRaster object from the terra package. Units: integers representing land cover classes. CRS: EPSG 27700 (British National Grid)
}
\source{
\url{https://EIDC/}
}
\usage{
get_r_lcm()
}
\description{
A raster layer of UK Land Cover Map classes. Object 'r_lcm' returned to global environment.
}
\examples{
get_r_lcm()
}
