% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spCEH.R
\name{get_r_twi}
\alias{get_r_twi}
\title{Function to export UK Topographic Wetness Index}
\format{
A SpatRaster object from the terra package. Units: dimensionless ratio. CRS: EPSG 27700 (British National Grid)
}
\source{
\url{https://Derived from OS DEM data by PL/}
}
\usage{
get_r_twi()
}
\description{
A raster layer of Topographic Wetness Index in the UK. Object 'r_twi' returned to global environment.
}
\examples{
get_r_twi()
}
