% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spCEH.R
\name{getSpatRasterTemplate}
\alias{getSpatRasterTemplate}
\title{Function to initialise an empty SpatRaster for
  the UK or a sub-region}
\usage{
getSpatRasterTemplate(domain = "UK", res = 100, proj = NULL)
}
\arguments{
\item{domain}{Domain of the output raster: "UK", "Scotland" etc.}

\item{res}{Resolution of the output raster in metres (for OSGB) or decimal degrees (WGS84). For the 'UK_NAME' domain, the res is pre-specified.}

\item{proj}{CRS (coordinate reference system) of the output raster: "OSGB" or "WGS84".}
}
\value{
An empty SpatRaster (from the terra package) object covering the domain.
}
\description{
Function to initialise an empty SpatRaster for
  the UK or a sub-region
}
\examples{
r <- getSpatRasterTemplate(domain = "UK", res = 10000, proj = 'OSGB')
r <- getSpatRasterTemplate(domain = "Scotland", res = 10000, proj = 'OSGB')
r <- getSpatRasterTemplate(domain = "NT_10km", res = 100, proj = 'OSGB')
r <- getSpatRasterTemplate(domain = "UK", res = 10000, proj = 'OSGB')
r <- getSpatRasterTemplate(domain = "UK", res = 0.1, proj = 'WGS84')
r <- getSpatRasterTemplate(domain = "UK_NAME", proj = 'WGS84')
r <- getSpatRasterTemplate(domain = "UK_NAME", proj = 'WGS84')
}
