% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spCEH.R
\docType{data}
\name{projIre}
\alias{projIre}
\title{CRS object for the TM75 Irish Grid projection.}
\format{
A text string
}
\source{
\url{https://spatialreference.org/}
}
\usage{
projIre
}
\description{
A Coordinate Reference System object for the Transverse Mercator projection used by the Ordnance Survey in Ireland (TM75 / EPSG:29903)
}
\keyword{datasets}
