% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spCEH.R
\docType{data}
\name{projlonlat}
\alias{projlonlat}
\title{CRS object for null projection (longitude-latitude).}
\format{
A text string
}
\source{
\url{https://spatialreference.org/}
}
\usage{
projlonlat
}
\description{
A Coordinate Reference System object for the WGS84 lon-lat coordinate system / EPSG:4326
}
\keyword{datasets}
