% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spCEH.R
\docType{data}
\name{projOSGB}
\alias{projOSGB}
\title{CRS object for the OSGB projection.}
\format{
A text string
}
\source{
\url{https://spatialreference.org/}
}
\usage{
projOSGB
}
\description{
A Coordinate Reference System object for the Transverse Mercator projection used by the Ordnance Survey in Great Britain (OSGB)
}
\keyword{datasets}
