% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spCEH.R
\docType{data}
\name{r_alt}
\alias{r_alt}
\title{UK altitude}
\format{
A Raster object from the raster package. Units: metres above mean sea level
}
\source{
\url{https://SRTM space shuttle terrain mission I think/}
}
\usage{
r_alt
}
\description{
A raster layer of altitude in the UK
}
\keyword{datasets}
