% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spCEH.R
\docType{data}
\name{r_Csoil}
\alias{r_Csoil}
\title{UK soil carbon}
\format{
A Raster object from the raster package. Units: kg C / m2
}
\source{
\url{https://Bradley et al 2005/}
}
\usage{
r_Csoil
}
\description{
A raster layer of soil carbon in the UK
}
\keyword{datasets}
