% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spCEH.R
\docType{data}
\name{r_lcm}
\alias{r_lcm}
\title{UK Land Cover Map classes}
\format{
A Raster object from the raster package. Units: integers representing land cover classes
}
\source{
\url{https://EIDC/}
}
\usage{
r_lcm
}
\description{
A raster layer of UK Land Cover Map classes
}
\keyword{datasets}
