% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spCEH.R
\docType{data}
\name{r_twi}
\alias{r_twi}
\title{UK Topographic Wetness Index}
\format{
A Raster object from the raster package. Units: dimensionless ratio
}
\source{
\url{https://Derived from OS DEM data by PL/}
}
\usage{
r_twi
}
\description{
A raster layer of Topographic Wetness Index in the UK
}
\keyword{datasets}
