% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spCEH.R
\docType{data}
\name{sfdf_uk}
\alias{sfdf_uk}
\title{UK coastline}
\format{
A SimpleFeaturesDataFrame object from the sf package
}
\source{
\url{https://spatialreference.org/}
}
\usage{
sfdf_uk
}
\description{
A SimpleFeaturesDataFrame of the UK coastline and borders
}
\keyword{datasets}
