% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spCEH.R
\docType{data}
\name{spgdf_uk}
\alias{spgdf_uk}
\title{UK coastline}
\format{
A SpatialPolygonsDataFrame object from the sp package
}
\source{
\url{https://spatialreference.org/}
}
\usage{
spgdf_uk
}
\description{
A SpatialPolygonsDataFrame of the UK coastline and borders
}
\keyword{datasets}
