% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spCEH.R
\name{maskByCountry}
\alias{maskByCountry}
\title{Function to mask out cells outwith polygons defining a country within the UK}
\usage{
maskByCountry(r, countryName)
}
\arguments{
\item{r}{A RasterLayer.}

\item{countryName}{The name of a country within the UK ("Scotland", "Northern Ireland", "England" or "Wales").}
}
\value{
A RasterLayer masked to the named country.
}
\description{
Function to mask out cells outwith polygons defining a country within the UK
}
\examples{
r <- getSpatRasterTemplate(domain = "UK", res = 10000, proj = 'OSGB')
r[] <- 1
r_masked <- maskByCountry(r, c("England", "Wales"))
plot(r_masked)
}
