% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spCEH.R
\docType{package}
\name{spCEH-package}
\alias{spCEH}
\alias{spCEH-package}
\title{Spatial functions and data from CEH.}
\description{
spCEH provides spatial utility functions and data.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/NERC-CEH/spCEH/issues}
}

}
\author{
\strong{Maintainer}: Peter Levy \email{plevy@ceh.ac.uk} (\href{https://orcid.org/0000-0002-8505-1901}{ORCID}) [copyright holder]

}
